[Feat] show my projects in homeView
1. 홈화면에서 내 프로젝트 뷰 노출 및 스크롤 가능하게 수정
