+++ teamplan/Sources/Services/Project/ProjectServicesCoredata.swift
@@ -187,8 +187,8 @@ extension ProjectServicesCoredata {
 
 
 // MARK: - Convert to DTO
-struct ProjectHomeDTO {
-    
+struct ProjectHomeDTO: Identifiable {
+    let id = UUID().uuidString
     let projectId: Int
     let title: String
     let finished: Bool
