+++ teamplan/Sources/Views/Home/SubView/MyProjectView.swift
@@ -13,7 +13,9 @@ struct MyProjectView: View {
     //MARK: Properties & Body
     @Binding var isProjectExist: Bool
     @State private var percent: CGFloat = 0.65
-    @EnvironmentObject private var homeVM: HomeViewModel
+    @State private var currentPage = 0
+
+    @ObservedObject var homeVM: HomeViewModel
     
     var body: some View {
         VStack{
@@ -61,132 +63,34 @@ struct MyProjectView: View {
     
     //MARK: Project List & CardView
     private var projectList: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: 16) {
-                ForEach(0..<3, id: \.self) { index in
-//                    let project = homeVM.userData.projects[index]
-//                    MyProjectCardView(stat: homeVM.statistics, project: project)
-//                        .frame(width: UIScreen.main.bounds.width - 32)
+        VStack {
+            TabView(selection: $currentPage) {
+                ForEach(Array(homeVM.userData.projects.enumerated()), id: \.element.id) { index, project in
+                    MyProjectCardView(stat: homeVM.userData.statData, project: project)
+                        .tag(index)
                 }
             }
-            .padding(.horizontal, 16)
-        }
-    }
-}
+            .frame(height: 194)
+            .tabViewStyle(PageTabViewStyle(indexDisplayMode: .never))
+            
+            pageControl
+                .padding(.top, 12)
 
-struct MyProjectCardView: View {
-    let stat: StatDTO
-    let project: ProjectHomeDTO
-    let percent: CGFloat = 0.65
-    
-    var body: some View {
-        VStack {
-            projectHead
-            projectTitle
-            projectTodoCount
-            Spacer().frame(height: 35)
-            projectProgressBar
-                .padding(.horizontal, 16)
-            projectDeadline
         }
-        .frame(height: 176)
-        .frame(maxWidth: .infinity)
-        .clipped()
-        .background(
-            Rectangle()
-                .foregroundColor(.clear)
-                .frame(width: 358, height: 176)
-                .background(.white)
-                .cornerRadius(8)
-                .shadow(color: .black.opacity(0.05), radius: 5, x: 0, y: 0)
-        )
+
+    }
+    
+    //MARK: Page Control
+    private var pageControl: some View {
+        HStack(spacing: 4) {
+            ForEach(0..<homeVM.userData.projects.count, id: \.self) { index in
+                Circle()
+                    .frame(width: 6, height: 6)
+                    .foregroundColor(index == currentPage ? .theme.mainPurpleColor : .init(hex: "D9D9D9"))
+            }
+        }
+        .frame(height: 6)
         .padding(.horizontal, 16)
-    }
-    
-    //MARK: Head
-    
-    private var projectHead: some  View {
-        HStack {
-            Text("마감기한이 다가오고 있어요!")
-                .foregroundColor(.theme.mainPurpleColor)
-                .font(.appleSDGothicNeo(.regular, size: 12))
-            Spacer()
-        }
-    }
-    
-    //MARK: Title
-    
-    private var projectTitle: some View {
-        HStack {
-            Text("\(project.title)")
-                .font(.appleSDGothicNeo(.bold, size: 18))
-                .foregroundColor(.black)
-            Spacer()
-            projectWaterCount
-        }
-    }
-    
-    //MARK: WaterDrop
-    
-    private var projectWaterCount: some View {
-        HStack {
-            Image("waterdrop")
-                .frame(width: 18, height: 14)
-            Text("\(stat.drop)")
-                .font(.appleSDGothicNeo(.semiBold, size: 18))
-            Text("개")
-                .font(.appleSDGothicNeo(.regular, size: 12))
-                .padding(.leading, -5)
-        }
-    }
-    
-    //MARK: Todo Count
-    
-    private var projectTodoCount: some View {
-        HStack {
-            Text("총 \(project.remainTodo)개의 TODO가 남아있어요")
-                .font(.appleSDGothicNeo(.regular, size: 12))
-                .foregroundColor(.black)
-            Spacer()
-        }
-    }
-    
-    //MARK: Progress Bar
-    /// ios 제공하는 기본 ProgressBar 사용으로 대체하였습니다.
-    /// ProgressView 는 주어진 값 (calcPercent()에 의해 계산된 값)과 총 값 (1.0)을 사용하여 진행 상태를 표시합니다.
-    ///  ProgressView의 스타일은 기존 둥근형태와 폭탄 아이콘 사용을 위해 CustomProgressViewStyle 을 사용합니다.
-    private var projectProgressBar: some View {
-        ProgressView(value: calcPercent(), total: 1.0)
-            .frame(height: 10)
-            .progressViewStyle(CustomProgressViewStyle())
-    }
-    
-    private func calcPercent() -> CGFloat {
-        let progress = CGFloat(project.progressedTerm) / CGFloat(project.totalTerm)
-        return min(max(progress, 0), 1)
-    }
-    
-    //MARK: DeadLine
-    
-    private var projectDeadline: some View {
-        VStack {
-            HStack {
-                Text("START")
-                    .font(.appleSDGothicNeo(.regular, size: 12))
-                    .foregroundColor(.theme.greyColor)
-                Spacer()
-                Text("D-\(project.remainDay)")
-                    .font(.appleSDGothicNeo(.regular, size: 12))
-                    .foregroundColor(.theme.blackColor)
-            }
-            Spacer().frame(height: 10)
-            HStack {
-                Spacer()
-                Text("\(DateFormatter.shortFormatter.string(from: project.startedAt)) - \(DateFormatter.shortFormatter.string(from: project.deadline))")
-                    .font(.appleSDGothicNeo(.regular, size: 12))
-                    .foregroundColor(.gray)
-            }
-        }
     }
 }
 
@@ -220,8 +124,8 @@ struct CustomProgressViewStyle: ProgressViewStyle {
 }
 
 
-struct MyProjectView_Previews: PreviewProvider {
-    static var previews: some View {
-        MyProjectView(isProjectExist: .constant(false))
-    }
-}
+//struct MyProjectView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        MyProjectView(isProjectExist: .constant(false))
+//    }
+//}
